fix: Use existing custom_shipstation_* field names
Reverted to using existing fields that already exist in database:
- custom_shipstation_tracking_number
- custom_shipstation_carrier
- custom_shipstation_shipping_cost

No need to rename fields - the 'shipstation' in the field name
is just internal. Users see the labels which can be anything.

This avoids:
- Creating duplicate fields
- Database migration
- Breaking existing DNs with old field names

The fields work for any shipping method (ShipStation, DHL, manual)
regardless of the internal field name.